% Generated by package preparation script.
\name{bvd_fill_ownership_panel}
\alias{bvd_fill_ownership_panel}
\title{Carry ownership values forward and backward within firm histories}
\description{Carry ownership values forward and backward within firm histories.}
\usage{
bvd_fill_ownership_panel(ownership, id_col = "ID_NUMBER", year_col = "YEAR", cols = NULL)
}
\arguments{
  \item{...}{See the source documentation and examples for argument details.}
}
\value{A data.table, vector, list, or other object as described in the function documentation.}
\details{This function is part of the R translation of the Kalemli-Özcan et al. (2024) Orbis data construction workflow.}
\references{Kalemli-Özcan, Ş., Sørensen, B. E., Villegas-Sanchez, C., Volosovych, V., and Yeşiltaş, S. (2024). American Economic Journal: Macroeconomics, 16(2), 353-374. doi:10.1257/mac.20220036.}
\examples{
# See README.md and the package vignettes for end-to-end examples.
}
