% Generated by package preparation script.
\name{bvd_size_distribution}
\alias{bvd_size_distribution}
\title{Compute firm size distribution by output, employment, and firm counts}
\description{Compute firm size distribution by output, employment, and firm counts.}
\usage{
bvd_size_distribution(x, by = c("CNTRYCDE", "YEAR"), employment_col = "EMPL", output_col = "OPER_TURN")
}
\arguments{
  \item{...}{See the source documentation and examples for argument details.}
}
\value{A data.table, vector, list, or other object as described in the function documentation.}
\details{This function is part of the R translation of the Kalemli-Özcan et al. (2024) Orbis data construction workflow.}
\references{Kalemli-Özcan, Ş., Sørensen, B. E., Villegas-Sanchez, C., Volosovych, V., and Yeşiltaş, S. (2024). American Economic Journal: Macroeconomics, 16(2), 353-374. doi:10.1257/mac.20220036.}
\examples{
# See README.md and the package vignettes for end-to-end examples.
}
