\name{orbiseu-package}
\alias{orbiseu-package}
\alias{orbiseu}
\docType{package}
\title{Make Orbis Firm-Level Data Representative for EU Economies}
\description{A data.table-based toolkit translating the Kalemli-Özcan et al. (2024) Orbis representativeness workflow into reusable R functions.}
\references{Kalemli-Özcan, Ş., Sørensen, B. E., Villegas-Sanchez, C., Volosovych, V., and Yeşiltaş, S. (2024). American Economic Journal: Macroeconomics, 16(2), 353-374. doi:10.1257/mac.20220036.}
